rto-web-app-main/rto-platform/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlalchemy==2.0.43
pandas==2.3.3
faker==37.8.0
scikit-learn==1.7.2
pytesseract==0.3.13
pillow==11.3.0
google-generativeai==0.8.5
python-multipart==0.0.20
pytest==8.4.2
sentence-transformers==2.2.2
